% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ArealAveraging}
\alias{ArealAveraging}
\title{Calculate the areal average based on a list of time series (e.g. from rain gauges)}
\usage{
ArealAveraging(tsList, areas = NULL, gisInfo = NULL,
  catchmentShape = NULL, interpolationMethod = "Thiessen")
}
\arguments{
\item{tsList}{list containing time series}

\item{areas}{percentage of total area covered by the Voronoi polygons related to each time series}

\item{gisInfo}{data.frame containing 6 variables: ID, STATION_NAME, BASIN_NAME,Type, X,Y, Elevation}

\item{catchmentShape}{SpatialPolygonsDataFrame containing the catchment boundary}

\item{interpolationMethod}{default value is "Thiessen". Available algorithms are: "Mean", "Thiessen",  "IDW" (Inverse Distance Weighted), "OK" (Ordinary Kriging)}
}
\value{
areal averaged time series
}
\description{
Calculate the areal average based on a list of time series (e.g. from rain gauges)
}
\author{
Claudia Vitolo
}

