% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalculateCN}
\alias{CalculateCN}
\title{Calculate the Curve Number from P and Q}
\usage{
CalculateCN(P, Q)
}
\arguments{
\item{P}{maximum precipitation for an event (it can also be a vector, for multiple events)}

\item{Q}{maximum discharge for an event (it can also be a vector, for multiple events)}
}
\value{
Curve Number, in the range [0,100]
}
\description{
Calculate the Curve Number from P and Q
}
\author{
Claudia Vitolo
}

