\name{catchmentShape}
\alias{catchmentShape}
\docType{data}
\title{
catchmentShape
}
\description{
catchmentShape is a spatial object that describes the catchment bounday. It can be used to test the function ArealInterpolation with the methods "IDW" and "OK".
}
\usage{data(catchmentShape)}
\format{
  The format is: SpatialPolygonsDataFrame
}
\examples{
data(catchmentShape)
## maybe str(catchmentShape) ; plot(catchmentShape) ...
}
\keyword{datasets}
