% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CommonRecordingTime}
\alias{CommonRecordingTime}
\title{Trim all the time series contained in a list to the common period.}
\usage{
CommonRecordingTime(tsList)
}
\arguments{
\item{tsList}{list of time series (e.g. zoo objects)}
}
\value{
list of time series with the same temporal window
}
\description{
Trim all the time series contained in a list to the common period.
}
\examples{
# CommonRecordingTime(tsList)
}
\author{
Claudia Vitolo
}

