% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CommonTemporalResolution}
\alias{CommonTemporalResolution}
\title{Find coarser temporal resolution amongst multiple time series}
\usage{
CommonTemporalResolution(myList)
}
\arguments{
\item{myList}{list containing multiple time series (e.g. zoo object)}
}
\value{
temporal resolution in seconds.
}
\description{
Find coarser temporal resolution amongst multiple time series
}
\details{
This function finds the coarser temporal resolution amongst a list of time series.
}
\examples{
# overlappingTS(myList)
}
\author{
Claudia Vitolo
}

