% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ConvertTS}
\alias{ConvertTS}
\title{This function converts inputs "from" one unit "to" another.}
\usage{
ConvertTS(dataset, from, to, optionalInput = NA)
}
\arguments{
\item{dataset}{Time series, it can be zoo, xts, etc}

\item{from}{string to specify the original unit, e.g. "m3/s" or "l/s"}

\item{to}{string to specify the destination unit, e.g. "mm/day" (default units to use with FUSE)}

\item{optionalInput}{real number which is used in case of special convertions (e.g. in case discharge is given in (from=) l/s and this needs to be converted to (to=)mm, then additional input is the area of the catchment)}
}
\value{
converted dataset
}
\description{
This function converts inputs "from" one unit "to" another.
}
\examples{
#ConvertTS( dataset, from, to, optionalInput=NA )
}
\author{
Claudia Vitolo
}
\references{
For the "from" and "to" strings, check the Unidata's udunits reference: http://www.unidata.ucar.edu/software/udunits/
}

