% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CorrectNeg}
\alias{CorrectNeg}
\title{Correct negative values.}
\usage{
CorrectNeg(tseries)
}
\arguments{
\item{tseries}{time series}
}
\value{
a time series with non negative values
}
\description{
Correct negative values.
}
\details{
At the moment the function simply substitutes negative values with NAs.
}
\examples{
# CorrectNeg( tseries )
}
\author{
Claudia Vitolo
}

