\name{DATA}
\alias{DATA}
\docType{data}
\title{
DATA
}
\description{
Data.frame containing forcing inputs (P, E and Q).
}
\usage{data("DATA")}
\format{
  The format is: zoo series from 2007-09-28 10:00:00 to 2009-12-01 11:00:00
}
\examples{
data(DATA)
## maybe str(DATA) ; plot(DATA) ...
}
\keyword{datasets}
