% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EventIdentification}
\alias{EventIdentification}
\title{Identify Rainfall-Runoff events}
\usage{
EventIdentification(data)
}
\arguments{
\item{data}{data.frame of precipitation (P) and discharge (Q) time series}
}
\value{
data.frame containing 3 columns: Tr (return period), P (max precipitation) and Q (max discharge)
}
\description{
Identify Rainfall-Runoff events
}
\author{
Claudia Vitolo
}

