% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractOverlappingPeriod}
\alias{ExtractOverlappingPeriod}
\title{This function trims P, E and Q time series to the overlapping period.
It is possible to ignore Q by setting ignoreQ=TRUE}
\usage{
ExtractOverlappingPeriod(P0, E0, Q0, ignoreQ = FALSE)
}
\arguments{
\item{P0}{precipitation time series (e.g. zoo object)}

\item{E0}{potential evapotranspiration time series (e.g. zoo object)}

\item{Q0}{streamflow discharge time series (e.g. zoo object)}

\item{ignoreQ}{boolean by default set to FALSE. Used to the algorithm to ignore Q.}
}
\value{
zoo object in which P, E and Q are merged.
}
\description{
This function trims P, E and Q time series to the overlapping period.
It is possible to ignore Q by setting ignoreQ=TRUE
}
\examples{
#ExtractOverlappingPeriod(P0,E0,Q0,ignoreQ=FALSE)
}
\author{
Claudia Vitolo
}

