% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fillGaps}
\alias{fillGaps}
\title{Infill missing values.}
\usage{
fillGaps(myList)
}
\arguments{
\item{myList}{List of datasets containing P, E and Q (with corrected recording times and no negative values for P and Q)}
}
\value{
a dataset with infilled missing values
}
\description{
Infill missing values.
}
\details{
It uses the fuction na.approx (better not use na.spline, as it generates negative values)
}
\examples{
# fillGaps( myList )
}
\author{
Claudia Vitolo
}

