% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findGaps}
\alias{findGaps}
\title{Finds gaps in time series}
\usage{
findGaps(myTS, deltim)
}
\arguments{
\item{myTS}{is a list of nested time series.}

\item{deltim}{final time step}
}
\value{
a list of fullranges, times and intervals of occurring gaps
}
\description{
Finds gaps in time series
}
\examples{
# findGaps( myTS, deltim )
}
\author{
Claudia Vitolo
}

