% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fullrangeTS}
\alias{fullrangeTS}
\title{Create a time series without gaps in recording}
\usage{
fullrangeTS(myTS, myRanges)
}
\arguments{
\item{myTS}{is a list of nested time series.}

\item{myRanges}{range obtained from function \code{findGaps}}
}
\value{
time series with no gaps in recording
}
\description{
Create a time series without gaps in recording
}
\examples{
# fullrangeTS( myTS, myRanges )
}

