% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glue}
\alias{glue}
\title{GLUE uncertainty analysis}
\usage{
glue(DATA, deltim, models, parameters, threshold)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects).
It is structured into three columns:
"P" containing precipitation,
"E" containing evapo-transpiration and
"Q" containing streamflow discharge.}

\item{deltim}{time step in days}

\item{models}{list of FUSE model structures to use}

\item{parameters}{table containing 24 parameters (columns) and a number of rows equal to the number of samples}

\item{threshold}{this is typically a number in the range [0.1,0.9]}
}
\value{
a plot showing uncertainties
}
\description{
GLUE uncertainty analysis
}
\examples{
# x <- glue(DATA, deltim, models, parameters, threshold)
}
\author{
Claudia Vitolo
}

