% Generated by roxygen2 (4.0.2): do not edit by hand
\name{irr2reg}
\alias{irr2reg}
\title{Transforms irregular time series to regular ones}
\usage{
irr2reg(myTS, deltim)
}
\arguments{
\item{myTS}{can only be a single time series}

\item{deltim}{final time step}
}
\value{
time series with regular recording time
}
\description{
Transforms irregular time series to regular ones
}
\details{
Sometimes recording time in datasets can be shifted by few minutes.
This function forces the minutes to be multiple of a certain "multiplier".
It works with a maximum of 2 levels of nested time series. Something like:
list("P"=list(P1,P2,P3), "E"=E, "Q"=Q)
}
\examples{
# irr2reg( myTS, deltim )
}

