% Generated by roxygen2 (4.0.2): do not edit by hand
\name{irreg2regTS}
\alias{irreg2regTS}
\title{Transforms irregular time series to regular ones}
\usage{
irreg2regTS(CatchmentName, DataList, deltim)
}
\arguments{
\item{CatchmentName}{string defining the name of the catchment}

\item{DataList}{number of minutes for the resulting regular time series. E.g. 15 (min)}

\item{deltim}{final time step}
}
\value{
time series with regular recording time
}
\description{
Transforms irregular time series to regular ones
}
\details{
Sometimes recording time in datasets can be shifted by few minutes.
This function forces the minutes to be multiple of a certain "multiplier".
It works only with PURE data with a maximum of 2 levels of nested time series. Something like: list("P"=list(P1,P2,P3), "E"=E, "Q"=Q)
}
\examples{
# irreg2regTS( CatchmentName, DataList, deltim )
}

