% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list2df}
\alias{list2df}
\title{Convert a list of time series (from rain gauges) to a data frame}
\usage{
list2df(P)
}
\arguments{
\item{P}{list of objects containing rainfall gauges time series}
}
\value{
data frame object
}
\description{
Convert a list of time series (from rain gauges) to a data frame
}
\details{
P should contain objects with same length, start and end date-time. Use function overlappingTS. See package hydroTSM for details on the x.ts and x.gis objects.
}
\examples{
# list2df(P)
}
\author{
Claudia Vitolo
}

