% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LoadMyData}
\alias{LoadMyData}
\title{Load data for any of the case studies (13 + 2).}
\usage{
LoadMyData(CatchmentName, SubcatchmentName, datafolder)
}
\arguments{
\item{CatchmentName}{this is a one-word string containing the name of the
       catchment. Accepted values are: "Eden" or "Pontbren"}

\item{SubcatchmentName}{this is an integer number representing the name of
       the subcatchment. Predefined values are:
       For Eden:      1 (Eden at BlindBeck),   2 (Eden at Dacre),
       For Pontbren:  1 (Pontbren at Site1),   2 (Pontbren at Site2),
                      3 (Pontbren at Site3),   4 (Pontbren at Site4),
                      5 (Pontbren at Site5),   6 (Pontbren at Site6),
                      7 (Pontbren at Site7),   8 (Pontbren at Site8),
                      9 (Pontbren at Site9),  10 (Pontbren at Site10),
                     11 (Pontbren at Site11), 12 (Pontbren at Site12),
                     13 (Pontbren at Site13)}

\item{datafolder}{path to the folder containing the data}
}
\value{
a list of 4 objects:
        $DATA a data.frame with 4 columns containing observations:
        date&time (Dates), discharge (Q), precipitation (P),
        potential evapo-transpiration (E)
        $deltimP observed precipitation time step
        $deltimE observed potential evapotranspiration time step
        $deltimQ observed streamflow discharge time step
}
\description{
Load data for any of the case studies (13 + 2).
}
\examples{
# LoadMyData(CatchmentName="Eden",SubcatchmentName=2, datafolder="/home/data/")
}

