% Generated by roxygen2 (4.0.2): do not edit by hand
\name{morecs_d2h}
\alias{morecs_d2h}
\title{This function imports evapotranspiration data from MORECS (daily) and interpolate to hourly resolution}
\usage{
morecs_d2h(filepath)
}
\arguments{
\item{filepath}{string representing the file location, e.g. "~/Dropbox/Data/pontbren/ET/DAILY L Vyrnwy.CSV"}
}
\value{
zoo object containing the hourly time series of potential evapotranspiration
}
\description{
This function imports evapotranspiration data from MORECS (daily) and interpolate to hourly resolution
}
\examples{
#morecs_d2h( filepath="~/Dropbox/Data/pontbren/ET/DAILY L Vyrnwy.CSV" )
}
\author{
Wouter Buytaert (2012/01/31), modified by Claudia Vitolo (2013/11/27)
}

