% Generated by roxygen2 (4.0.2): do not edit by hand
\name{overlappingTS}
\alias{overlappingTS}
\title{This function trims a list made of multiple datasets (e.g. multiple rain gauges datsets) to the overlapping period.}
\usage{
overlappingTS(myList)
}
\arguments{
\item{myList}{list containing multiple time series (e.g. zoo object)}
}
\value{
list of zoo objects in which each time series is defined for the same overlapping period.
}
\description{
This function trims a list made of multiple datasets (e.g. multiple rain gauges datsets) to the overlapping period.
}
\examples{
# overlappingTS(myList)
}
\author{
Claudia Vitolo
}

