% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pet}
\alias{pet}
\title{Calculate Potential Evapotranspiration using Penman Monteith formula}
\usage{
pet(stationElevation = 0, TD, TW, NR, WS)
}
\arguments{
\item{stationElevation}{elevation of the recording station (in metres)}

\item{TD}{time series of dry bulb temperature}

\item{TW}{time series of wet bulb temperature}

\item{NR}{time series of net radiation}

\item{WS}{time series of wind speed}
}
\value{
ET0 = ET for grass (K = 1)
}
\description{
Calculate Potential Evapotranspiration using Penman Monteith formula
}
\examples{
# data(exampleTimeSeries)
# pet(stationElevation=0,TD,TW,NR,WS)
}
\author{
Claudia Vitolo
}
\references{
FAO - hourly method: http://www.fao.org/docrep/x0490e/x0490e08.htm#TopOfPage
}

