% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RainfallArealAveraging}
\alias{RainfallArealAveraging}
\title{Function to calculate the rainfall areal average based on a list of time series (from rain gauges)}
\usage{
RainfallArealAveraging(CatchmentName, Plist, DataList,
  InterpolationMethod = "Thiessen")
}
\arguments{
\item{CatchmentName}{string identifying the name of the catchment}

\item{Plist}{list containing rainfall gauges time series}

\item{DataList}{list of objects from function LoadMyData}

\item{InterpolationMethod}{default value is "Thiessen". Available algorithms are: "AritmeticMean", "Thiessen",  "IDW" (Inverse Distance Weighted), "OK" (Ordinary Kriging)}
}
\value{
areal averaged time series
}
\description{
Function to calculate the rainfall areal average based on a list of time series (from rain gauges)
}
\author{
Claudia Vitolo
}

