% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scanTimeSeries}
\alias{scanTimeSeries}
\title{Scans time series in a named list}
\usage{
scanTimeSeries(myTS)
}
\arguments{
\item{myTS}{list of time series, it can be zoo, xts, etc}
}
\value{
print useful information
}
\description{
Scans time series in a named list
}
\examples{
# myTS( dataset=Q[[8]])
}

