% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CommonTemporalResolution}
\alias{CommonTemporalResolution}
\title{Use a common temporal resolution for multiple time series}
\usage{
CommonTemporalResolution(myList)
}
\arguments{
\item{myList}{list containing multiple time series (e.g. zoo object)}
}
\value{
list of zoo objects with the same temporal resolution.
}
\description{
Use a common temporal resolution for multiple time series
}
\details{
This function finds the coarser temporal resolution amongst a list of time series an aggregates them all.
}
\examples{
# overlappingTS(myList)
}
\author{
Claudia Vitolo
}

