% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correctNeg}
\alias{correctNeg}
\title{Function to correct negative values.}
\usage{
correctNeg(regTSList)
}
\arguments{
\item{regTSList}{List of datasets containing P, E and Q (with corrected recording times)}
}
\value{
a dataset with non negative values
}
\description{
Function to correct negative values.
}
\details{
At the moment the function simply substitutes negative values with NAs.
}
\examples{
# correctNeg( regTSList )
}

