% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractOverlappingPeriod}
\alias{ExtractOverlappingPeriod}
\title{This function trims P, E and Q time series to the overlapping period.
It is possible to ignore Q by setting ignoreQ=TRUE}
\usage{
ExtractOverlappingPeriod(tsList)
}
\arguments{
\item{tsList}{list of time series (e.g. zoo objects)}
}
\value{
list of time series with the same temporal window
}
\description{
This function trims P, E and Q time series to the overlapping period.
It is possible to ignore Q by setting ignoreQ=TRUE
}
\examples{
# ExtractOverlappingPeriod(tsList)
}
\author{
Claudia Vitolo
}

