% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Irr2Reg}
\alias{Irr2Reg}
\title{Transforms irregular time series to regular ones}
\usage{
Irr2Reg(dataset)
}
\arguments{
\item{dataset}{can only be a single time series}
}
\value{
time series with regular recording time
}
\description{
Transforms irregular time series to regular ones
}
\details{
Sometimes recording time in datasets can be shifted by few minutes.
This function forces the minutes to be multiple of a certain "multiplier".
}
\examples{
# Irr2Reg( dataset )
}
\author{
Claudia Vitolo
}

