% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MCsimulations}
\alias{MCsimulations}
\title{This function runs Monte Carlo simulations for a given number of realisations.}
\usage{
MCsimulations(DATA, deltim, warmup, parameters, ModelList, SimulationFolder,
  MPIs, verbose = TRUE)
}
\arguments{
\item{DATA}{this is the data frame containing the observations}

\item{deltim}{this is the time step (in days)}

\item{warmup}{this is the warmup period}

\item{parameters}{table containing the parameter sets, see \code{GeneratePsetsFUSE}}

\item{ModelList}{this is the reduced list of model structures (total number: 240)}

\item{SimulationFolder}{path to the folder where the function saves the results}

\item{MPIs}{list of functions describing the Model performance Indices}

\item{verbose}{if TRUE, it prints details of the current simulation}
}
\value{
2 arrays: R (discharges) and S (indices).
}
\description{
This function runs Monte Carlo simulations for a given number of realisations.
}
\examples{
# MCsimulations(DATA,deltim,warmup,parameters,ModelList,SimulationFolder,MPIs)
}

