% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ScanTS}
\alias{ScanTS}
\title{Scan time series given in input and check for un-even recording times,
negative values and missing values.}
\usage{
ScanTS(dataset, returnNegInfo = TRUE, returnTimeInfo = TRUE,
  returnMissingsInfo = TRUE, verbose = FALSE)
}
\arguments{
\item{dataset}{Time series, it can be zoo, xts, etc}

\item{returnNegInfo}{bolean, default value is TRUE}

\item{returnTimeInfo}{bolean, default value is TRUE}

\item{returnMissingsInfo}{bolean, default value is TRUE}

\item{verbose}{bolean, default value is TRUE}
}
\value{
print useful information and suggestions
}
\description{
Scan time series given in input and check for un-even recording times,
negative values and missing values.
}
\examples{
# ScanTS( dataset )
}
\author{
Claudia Vitolo
}

